moved static pcaps to resources, moved vector code to ssh.h
