+++ src/ssh.c
@@ -53,8 +53,143 @@
 
 #include <string.h>     /* for memset()            */
 #include "ssh.h"
+#include "utils.h"      /* for enum role */
 #include "p2f.h"        /* for zprintf_ ...        */
 #include "err.h"        /* for logging             */
+
+/*
+ *
+ * \brief Copy a json-printable string from the source buffer to the
+ *        destination buffer.
+ *
+ * \param buf Destination buffer.
+ * \param buflen Maximum length of the destination buffer.
+ * \param data Source buffer.
+ * \param datalen Length of the source buffer.
+ *
+ */
+void copy_printable_string(char *buf,
+			   unsigned int buflen,
+			   const void *data,
+			   unsigned int datalen) {
+    const char *d = data;
+
+    while (buflen-- && datalen--) {
+	if (!isprint(*d) || *d == '\"' || *d == '\\' || *d <= 0x1f) { /* json constraints */
+	    break;
+	}
+	*buf++ = *d++;
+    }
+
+    *buf = 0; /* null terminate buffer */
+}
+
+/*
+ *
+ * \brief Initialize vector struct.
+ *
+ * \param vector Pointer to the vector to be initialized.
+ *
+ */
+void vector_init(struct vector *vector) {
+
+    vector->len = 0;
+    vector->bytes = NULL;
+
+    return;
+}
+
+/*
+ *
+ * \brief Set the vector contents to the specified data, freeing the previous
+ * vector contents. If the previous vector contents overlap in memory with the
+ * new vector contents, the behavior is still defined since the free occurs
+ * after the copy.
+ *
+ * \param vector Pointer to the vector to be set.
+ * \param data Pointer to byte array to be copied.
+ * \param len Length of the byte array to be copied.
+ *
+ */
+void vector_set(struct vector *vector, const void *data, unsigned len) {
+    void *tmpptr = NULL;
+
+    tmpptr = malloc(len);
+    if (tmpptr == NULL) {
+        return;
+    }
+    memcpy(tmpptr, data, len);
+    vector_free(vector); /* does nothing if already empty */
+    vector->bytes = tmpptr;
+    vector->len = len;
+
+    return;
+}
+
+/*
+ *
+ * \brief Append the specified data to the current vector contents, even if the
+ * vector is currently empty.
+ *
+ * \param vector Pointer to the vector to be appended to.
+ * \param data Pointer to byte array to be appended.
+ * \param len Length of the byte array to be appended.
+ *
+ */
+void vector_append(struct vector *vector, const void *data, unsigned len) {
+
+    vector->bytes = realloc(vector->bytes, vector->len + len);
+    if (vector->bytes == NULL) {
+        return;
+    }
+    memcpy(vector->bytes + vector->len, data, len);
+    vector->len += len;
+
+    return;
+}
+
+/*
+ *
+ * \brief Allocate and return a pointer to a string representation of a vector.
+ * This string must later be freed to avoid a memory leak.
+ *
+ * \param vector Pointer to the vector to stringify.
+ *
+ * \return A pointer to a string representation of the vector.
+ */
+char *vector_string(struct vector *vector) {
+    char *s;
+
+    s = malloc(vector->len+1);
+    if (s == NULL) {
+        return NULL;
+    }
+    if (vector->len > 0) {
+        copy_printable_string(s, vector->len+1, vector->bytes, vector->len);
+    } else {
+        s[0] = 0;
+    }
+
+    return s;
+}
+
+/*
+ *
+ * \brief Free a vector, setting it to the state just after a call to vector_init.
+ *
+ * \param vector Pointer to the vector to free.
+ *
+ */
+void vector_free(struct vector *vector) {
+
+    if (vector->bytes != NULL) {
+        free(vector->bytes);
+        vector->bytes = NULL;
+    }
+    vector->len = 0;
+
+    return;
+}
 
 /*
  * from http://www.iana.org/assignments/ssh-parameters/ssh-parameters.xhtml
