+++ src/utils.c
@@ -203,94 +203,3 @@ JSON_Value* joy_utils_open_resource_parson(const char *filename) {
 
     return value;
 }
-
-/*
- *
- * \brief Copy a json-printable string from the source buffer to the
- *        destination buffer.
- *
- * \param buf Destination buffer.
- * \param buflen Maximum length of the destination buffer.
- * \param data Source buffer.
- * \param datalen Length of the source buffer.
- *
- */
-void copy_printable_string(char *buf,
-			   unsigned int buflen,
-			   const void *data,
-			   unsigned int datalen) {
-    const char *d = data;
-
-    while (buflen-- && datalen--) {
-	if (!isprint(*d) || *d == '\"' || *d == '\\' || *d <= 0x1f) { /* json constraints */
-	    break;
-	}
-	*buf++ = *d++;
-    }
-
-    *buf = 0; /* null terminate buffer */
-}
-
-void vector_init(struct vector *vector) {
-
-    vector->len = 0;
-    vector->bytes = NULL;
-
-    return;
-}
-
-/* set vector to data, where data may overlap with current vector contents */
-void vector_set(struct vector *vector, const void *data, unsigned len) {
-    void *tmpptr = NULL;
-
-    tmpptr = malloc(len);
-    if (tmpptr == NULL) {
-        return;
-    }
-    memcpy(tmpptr, data, len);
-    vector_free(vector); /* does nothing if already empty */
-    vector->bytes = tmpptr;
-    vector->len = len;
-
-    return;
-}
-
-/* append data to current vector contents (even if vector is empty) */
-void vector_append(struct vector *vector, const void *data, unsigned len) {
-
-    vector->bytes = realloc(vector->bytes, vector->len + len);
-    if (vector->bytes == NULL) {
-        return;
-    }
-    memcpy(vector->bytes + vector->len, data, len);
-    vector->len += len;
-
-    return;
-}
-
-char *vector_string(struct vector *vector) {
-    char *s;
-
-    s = malloc(vector->len+1);
-    if (s == NULL) {
-        return NULL;
-    }
-    if (vector->len > 0) {
-        copy_printable_string(s, vector->len+1, vector->bytes, vector->len);
-    } else {
-        s[0] = 0;
-    }
-
-    return s;
-}
-
-void vector_free(struct vector *vector) {
-
-    if (vector->bytes != NULL) {
-        free(vector->bytes);
-        vector->bytes = NULL;
-    }
-    vector->len = 0;
-
-    return;
-}
